# State-machine-ify benchmark.zig.