Fixes https://github.com/tigerbeetledb/tigerbeetle/issues/588.

```
# before
> ./scripts/benchmark.sh --transfer-count 1000000 --transfer-count-per-second 10000
load offered = 10000 tx/s
load accepted = 9990 tx/s
batch latency p00 = 5 ms
batch latency p10 = 5 ms
batch latency p20 = 5 ms
batch latency p30 = 5 ms
batch latency p40 = 5 ms
batch latency p50 = 5 ms
batch latency p60 = 5 ms
batch latency p70 = 5 ms
batch latency p80 = 5 ms
batch latency p90 = 5 ms
batch latency p100 = 858 ms
transfer latency p00 = 5 ms
transfer latency p10 = 6 ms
transfer latency p20 = 7 ms
transfer latency p30 = 7 ms
transfer latency p40 = 8 ms
transfer latency p50 = 9 ms
transfer latency p60 = 17 ms
transfer latency p70 = 73 ms
transfer latency p80 = 156 ms
transfer latency p90 = 273 ms
transfer latency p100 = 893 ms

# after
> ./scripts/benchmark.sh --transfer-count 1000000 --transfer-count-per-second 10000
25983 batches in 100.06 s
load offered = 10000 tx/s
load accepted = 9994 tx/s
batch latency p00 = 0 ms
batch latency p10 = 0 ms
batch latency p20 = 0 ms
batch latency p30 = 0 ms
batch latency p40 = 0 ms
batch latency p50 = 1 ms
batch latency p60 = 1 ms
batch latency p70 = 1 ms
batch latency p80 = 1 ms
batch latency p90 = 2 ms
batch latency p100 = 882 ms
transfer latency p00 = 0 ms
transfer latency p10 = 1 ms
transfer latency p20 = 2 ms
transfer latency p30 = 4 ms
transfer latency p40 = 26 ms
transfer latency p50 = 61 ms
transfer latency p60 = 106 ms
transfer latency p70 = 161 ms
transfer latency p80 = 230 ms
transfer latency p90 = 327 ms
transfer latency p100 = 915 ms
```